Fix compiler warning, update version

+++ engine/sc_util.cpp
@@ -2608,7 +2608,7 @@ std::string util::create_blizzard_talent_url( const player_t* p )
    default: assert( 0 ); break;
   }
   url += "!";
-  for ( size_t i = 0; i < MAX_TALENT_ROWS; i++ )
+  for ( int i = 0; i < MAX_TALENT_ROWS; i++ )
   {
     if ( p -> talent_points.choice( i ) >= 0 )
       url += util::to_string( p -> talent_points.choice( i ) );
